End byte is implemented!
It has been a long road but I finally got it... There was really weird behavior pulling a single byte out of the TCP buffer at a time, but it seems the "readStringUntil" method works pretty well to read until the null char. I also came to the conclusion that the packets won't come in a nice rythm making the animation stuttery. I implemented an nice little buffer with drift detection to solve this. It tries to keep the buffer only 5 to 10 frames long.